# src/InactivePDF.Api/wwwroot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InactivePDF API",
    "version": "0.1.0-beta",
    "description": "Self-hosted document-to-PDF conversion API. Use an administrator bearer token for administration routes or a scoped integration key for conversion and owned-job routes."
  },
  "servers": [{ "url": "/", "description": "Current InactivePDF instance" }],
...
  "paths": {
...
    "/v1/jobs": {
      "post": { "tags": ["Jobs"], "summary": "Submit an asynchronous job", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/JobMultipart" }, "responses": { "202": { "$ref": "#/components/responses/AcceptedJob" }, "default": { "$ref": "#/components/responses/Problem" } } },
      "get": { "tags": ["Jobs"], "summary": "List owned jobs", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/State" }, { "$ref": "#/components/parameters/Format" }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "A cursor-paginated job list.", "content": { "application/json": { "schema": { "type": "object" } } } }, "default": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/v1/profiles": { "get": { "tags": ["Profiles"], "summary": "List output and watermark profiles", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "responses": { "200": { "description": "Profile catalog", "content": { "application/json": { "schema": { "type": "object" } } } } } } },
    "/v1/profiles/{name}": { "get": { "tags": ["Profiles"], "summary": "Look up a named profile", "security": [{ "integrationBearer": [] }, { "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Profile definition" }, "404": { "$ref": "#/components/responses/Problem" } } } },
...
    "/v1/admin/api-keys": {
      "get": { "tags": ["Administration"], "summary": "List API-key metadata", "security": [{ "administratorBearer": [] }], "responses": { "200": { "description": "Key metadata. Secrets are never returned." } } },
      "post": { "tags": ["Administration"], "summary": "Create an integration API key", "security": [{ "administratorBearer": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiKeyCreateRequest" } } } }, "responses": { "201": { "description": "Metadata and the one-time secret" }, "default": { "$ref": "#/components/responses/Problem" } } }
    },
    "/v1/admin/api-keys/{id}": { "get": { "tags": ["Administration"], "summary": "Read API-key metadata", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Problem" } } }, "delete": { "tags": ["Administration"], "summary": "Delete an API key", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "204": { "description": "Deleted" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/api-keys/{id}/rotate": { "post": { "tags": ["Administration"], "summary": "Rotate an API key", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/api-keys/{id}/revoke": { "post": { "tags": ["Administration"], "summary": "Revoke an API key", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/api-keys/{id}/limits": { "put": { "tags": ["Administration"], "summary": "Update API-key limits", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } }, "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/api-keys/{id}/usage": { "get": { "tags": ["Administration"], "summary": "Get usage for one API key", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/profiles": { "post": { "tags": ["Administration"], "summary": "Create a watermark profile", "security": [{ "administratorBearer": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileWriteRequest" } } } }, "responses": { "201": { "$ref": "#/components/responses/Json" }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/profiles/{name}": { "put": { "tags": ["Administration"], "summary": "Update a watermark profile", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileWriteRequest" } } } }, "responses": { "200": { "$ref": "#/components/responses/Json" }, "400": { "$ref": "#/components/responses/Problem" } } }, "delete": { "tags": ["Administration"], "summary": "Delete a watermark profile", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["watermark"] } }], "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/performance-presets": { "get": { "tags": ["Administration"], "summary": "List built-in performance presets", "security": [{ "administratorBearer": [] }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/service-profiles": { "get": { "tags": ["Administration"], "summary": "List saved service profiles", "security": [{ "administratorBearer": [] }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/service-profiles/{name}": { "get": { "tags": ["Administration"], "summary": "Read a saved service profile", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "404": { "$ref": "#/components/responses/Problem" } } }, "put": { "tags": ["Administration"], "summary": "Replace a saved service profile", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } }, "responses": { "204": { "description": "Saved" }, "400": { "$ref": "#/components/responses/Problem" } } }, "delete": { "tags": ["Administration"], "summary": "Delete a saved service profile", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/usage": { "get": { "tags": ["Administration"], "summary": "Get all API-key usage and totals", "security": [{ "administratorBearer": [] }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/audit-events": { "get": { "tags": ["Administration"], "summary": "Read safe audit events", "security": [{ "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/settings": { "get": { "tags": ["Administration"], "summary": "Read effective settings", "security": [{ "administratorBearer": [] }], "responses": { "200": { "$ref": "#/components/responses/Json" } } }, "put": { "tags": ["Administration"], "summary": "Save settings with optimistic concurrency", "security": [{ "administratorBearer": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["settings", "revision"] } } } }, "responses": { "200": { "$ref": "#/components/responses/Json" }, "409": { "$ref": "#/components/responses/Problem" }, "default": { "$ref": "#/components/responses/Problem" } } } },
    "/v1/admin/status": { "get": { "tags": ["Administration"], "summary": "Read health and runtime status", "security": [{ "administratorBearer": [] }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/jobs": { "get": { "tags": ["Administration"], "summary": "List recent jobs for operators", "security": [{ "administratorBearer": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/logs": { "get": { "tags": ["Administration"], "summary": "Read structured operator logs", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "q", "in": "query", "schema": { "type": "string", "maxLength": 128 } }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
...
    "/v1/admin/analytics": { "get": { "tags": ["Administration"], "summary": "Read conversion analytics", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "hours", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 168, "default": 24 } }, { "name": "source", "in": "query", "schema": { "type": "string", "enum": ["WatchFolder", "Synchronous", "Queued"] } }], "responses": { "200": { "$ref": "#/components/responses/Json" } } } },
    "/v1/admin/search": { "get": { "tags": ["Administration"], "summary": "Search operator metadata", "security": [{ "administratorBearer": [] }], "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 2, "maxLength": 128 } }, { "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Problem" } } } }
  },
  "components": {
    "securitySchemes": {
      "administratorBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "administrator token" },
      "integrationBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "integration API key" }
    },
    "parameters": {
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "schema": { "type": "string", "maxLength": 128 }, "description": "Repeat-safe identity for asynchronous submissions." },
      "State": { "name": "state", "in": "query", "schema": { "type": "string" } },
      "Format": { "name": "format", "in": "query", "schema": { "type": "string" } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } }
    },
    "requestBodies": {
...
      "JobMultipart": { "required": true, "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/JobForm" } } } }
    },
    "schemas": {
      "WatermarkOptions": { "type": "object", "properties": { "kind": { "type": "string", "enum": ["Text", "Image"], "default": "Text" }, "text": { "type": "string" }, "imagePath": { "type": "string", "description": "Protected asset filename, never an arbitrary server path." }, "fontFamily": { "type": "string", "default": "Arial" }, "fontSize": { "type": "number", "format": "double", "minimum": 1 }, "color": { "type": "string", "example": "#808080" }, "opacity": { "type": "number", "format": "double", "minimum": 0, "maximum": 1 }, "rotation": { "type": "number", "format": "double", "minimum": -360, "maximum": 360 }, "position": { "type": "string", "enum": ["TopLeft", "TopCenter", "TopRight", "CenterLeft", "Center", "CenterRight", "BottomLeft", "BottomCenter", "BottomRight"] }, "offsetX": { "type": "number", "format": "double" }, "offsetY": { "type": "number", "format": "double" }, "layer": { "type": "string", "enum": ["Behind", "Over"] }, "pages": { "type": "string", "example": "all" }, "tile": { "type": "boolean" }, "width": { "type": "number", "format": "double", "exclusiveMinimum": 0 }, "height": { "type": "number", "format": "double", "exclusiveMinimum": 0 }, "header": { "type": "string", "nullable": true }, "footer": { "type": "string", "nullable": true }, "pageNumberFormat": { "type": "string", "default": "{page}/{pages}" } }, "description": "WatermarkOptions. When only one image dimension is supplied, the other is derived proportionally." },
      "ProfileWriteRequest": { "type": "object", "required": ["type", "options"], "properties": { "name": { "type": "string" }, "type": { "type": "string", "enum": ["watermark"] }, "options": { "$ref": "#/components/schemas/WatermarkOptions" } } },
      "ConversionForm": { "type": "object", "required": ["file"], "properties": { "mode": { "type": "string", "enum": ["sync", "async"], "default": "async" }, "profile": { "type": "string", "default": "archive" }, "watermarkProfile": { "type": "string" }, "watermark": { "type": "string", "description": "JSON-encoded WatermarkOptions for multipart requests. Direct options take precedence over watermarkProfile." }, "file": { "type": "array", "items": { "type": "string", "format": "binary" } } } },
      "JobForm": { "allOf": [{ "$ref": "#/components/schemas/ConversionForm" }, { "type": "object", "properties": { "operation": { "type": "string", "enum": ["ConvertFile", "ConvertAndMerge", "CreateTextPdf"] }, "correlationId": { "type": "string" }, "text": { "type": "string" } } }] },
...
      "ApiKeyCreateRequest": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 64 }, "scopes": { "type": "array", "items": { "type": "string" } }, "expiresAt": { "type": "string", "format": "date-time" }, "limits": { "type": "object" } } },
      "JobStatus": { "type": "object", "properties": { "jobId": { "type": "string", "format": "uuid" }, "state": { "type": "string" }, "progressPercent": { "type": "integer" }, "queueWaitMilliseconds": { "type": "integer" }, "processingMilliseconds": { "type": "integer" }, "retryCount": { "type": "integer" }, "inputBytes": { "type": "integer", "format": "int64" }, "outputBytes": { "type": "integer", "format": "int64" }, "compressionPercent": { "type": "number", "format": "double" } } },
      "Problem": { "type": "object", "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" } } }
    },
    "responses": {
...
      "AcceptedJob": { "description": "Job accepted", "headers": { "Location": { "schema": { "type": "string" } }, "X-Request-ID": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } },
      "Problem": { "description": "Structured error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Json": { "description": "JSON response", "content": { "application/json": { "schema": { "type": "object" } } } }
    }
  }
}